pyproject.toml
[tool.poetry]
name = "research_assistant"
version = "0.1.0"
description = "Research Agent PoC with SmoLAgents"

[tool.poetry.dependencies]
python = "^3.9"
smolagents = "^0.1.0"
docling = "^0.1.0"
llama-index = "^0.9.0"
streamlit = "^1.41.0"
opentelemetry-api = "^1.23.0"
opentelemetry-sdk = "^1.23.0"
opentelemetry-exporter-otlp = "^1.23.0"
structlog = "^24.1.0"
pydantic = "^2.0.0"